requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
tenacity==8.2.3
pandas==2.1.4
numpy==1.26.2
jinja2==3.1.2
